--- api/test/contracts/Flipper/abi.ts
@@ -1,35 +1,0 @@
-import { ABI } from "../../../src/types";
-
-export default [
-  {
-    "inputs": [
-      {
-        "internalType": "bool",
-        "name": "initvalue",
-        "type": "bool"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "constructor"
-  },
-  {
-    "inputs": [],
-    "name": "flip",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "get",
-    "outputs": [
-      {
-        "internalType": "bool",
-        "name": "",
-        "type": "bool"
-      }
-    ],
-    "stateMutability": "view",
-    "type": "function"
-  }
-] as ABI;
