--- api/test/contracts/Flipper/bytecode.ts
@@ -1,2 +1,0 @@
-export default "0x608060405234801561001057600080fd5b5060405161011938038061011983398101604081905261002f91610045565b6000805460ff191691151591909117905561006c565b600060208284031215610056578081fd5b81518015158114610065578182fd5b9392505050565b609f8061007a6000396000f3fe6080604052348015600f57600080fd5b506004361060325760003560e01c80636d4ce63c146037578063cde4efa9146051575b600080fd5b60005460ff16604051901515815260200160405180910390f35b60676000805460ff19811660ff90911615179055565b00fea264697066735822122088b93d28ee538a7f1b387b6b0dd71e85bbd61d827f2e1f588dc6aa88b86177e464736f6c634300080400330000000000000000000000000000000000000000000000000000000000000001";
-// args: true
