--- api/test/contracts/Flipper/source.ts
@@ -1,23 +1,0 @@
-export default JSON.stringify({
-  "Flipper.sol": `
-pragma solidity >=0.7.0 <0.9.0;
-
-contract Flipper {
-  bool private value;
-  /// Constructor that initializes the bool value to the given init_value.
-  constructor(bool initvalue) public {
-      value = initvalue;
-  }
-  /// A message that can be called on instantiated contracts.
-  /// This one flips the value of the stored bool from true
-  /// to false and vice versa.
-  function flip() public {
-      value = !value;
-  }
-  /// Simply returns the current value of our bool.
-  function get() public view returns (bool) {
-      return value;
-  }
-}  
-`.trim()
-});
