--- api/test/contracts/Greete/source.ts
@@ -1,22 +1,0 @@
-export default JSON.stringify({
-  'Greete.sol': `
-//SPDX-License-Identifier: Unlicense
-pragma solidity ^0.8.4;
-
-contract Greeter {
-  string greeting;
-
-  constructor(string memory _greeting) {
-    greeting = _greeting;
-  }
-
-  function greet() public view returns (string memory) {
-    return greeting;
-  }
-
-  function setGreeting(string memory _greeting) public {
-    greeting = _greeting;
-  }
-}
-`
-})
