--- api/test/contracts/Storage/bytecode.ts
@@ -1,2 +1,0 @@
-export default "0x608060405234801561001057600080fd5b5060405161010838038061010883398101604081905261002f91610037565b60005561004f565b600060208284031215610048578081fd5b5051919050565b60ab8061005d6000396000f3fe6080604052348015600f57600080fd5b506004361060325760003560e01c80632e64cec11460375780636057361d14604c575b600080fd5b60005460405190815260200160405180910390f35b605c6057366004605e565b600055565b005b600060208284031215606e578081fd5b503591905056fea2646970667358221220eefd54f067e6c9a97316f0b8ed80de98cec9b725b67b17c2d717155e275df51164736f6c63430008040033000000000000000000000000000000000000000000000000000000000096bfa5";
-// args 9879461
