--- api/test/contracts/Storage/source.ts
@@ -1,35 +1,0 @@
-export default JSON.stringify({
-  'Storage.sol': `
-// SPDX-License-Identifier: GPL-3.0
-
-pragma solidity >=0.7.0 <0.9.0;
-
-/**
- * @title Storage
- * @dev Store & retrieve value in a variable
- */
-contract Storage {
-
-    uint256 number;
-    
-    constructor(uint256 value) {
-        number = value;
-    }
-    /**
-     * @dev Store value in variable
-     * @param num value to store
-     */
-    function store(uint256 num) public {
-        number = num;
-    }
-
-    /**
-     * @dev Return value 
-     * @return value of 'number'
-     */
-    function retrieve() public view returns (uint256){
-        return number;
-    }
-}
-`
-})
